UniversalAdvocate_Submission/requirements.txt
google-generativeai
streamlit
python-dotenv
googlesearch-python
